+++ src/ast/lists/expressionList.cpp
@@ -3,7 +3,7 @@
 void ExpressionList::printC(std::ostream &os) const
 {
     std::vector<StatementPtr> initializations = getNodes();
-    for (int i = 0; i < initializations.size(); ++i)
+    for (size_t i = 0; i < initializations.size(); ++i)
     {
         if (i > 0) { os << ", "; }
         os << initializations[i];
@@ -13,7 +13,7 @@ void ExpressionList::printC(std::ostream &os) const
 void ExpressionList::generatePython(std::ostream &os, PythonContext &context, int scopeDepth) const
 {
     std::vector<StatementPtr> initializations = getNodes();
-    for (int i = 0; i < initializations.size(); ++i)
+    for (size_t i = 0; i < initializations.size(); ++i)
     {
         if (i > 0) { os << ", "; }
         initializations[i]->generatePython(os, context, scopeDepth);
