+++ src/ast/lists/parameterList.cpp
@@ -3,7 +3,7 @@
 void ParameterList::printC(std::ostream &os) const
 {
     std::vector<StatementPtr> initializations = getNodes();
-    for (int i = 0; i < initializations.size(); ++i)
+    for (size_t i = 0; i < initializations.size(); ++i)
     {
         if (i > 0) { os << ", "; }
         os << initializations[i];
@@ -13,7 +13,7 @@ void ParameterList::printC(std::ostream &os) const
 void ParameterList::generatePython(std::ostream &os, PythonContext &context, int scopeDepth) const
 {
     std::vector<StatementPtr> initializations = getNodes();
-    for (int i = 0; i < initializations.size(); ++i)
+    for (size_t i = 0; i < initializations.size(); ++i)
     {
         if (i > 0) { os << ", "; }
         initializations[i]->generatePython(os, context, scopeDepth);
@@ -22,7 +22,7 @@ void ParameterList::generatePython(std::ostream &os, PythonContext &context, int
 
 void ParameterList::generateIL(std::vector<Instr> &instrs, ILContext &context, std::string destReg) const
 {
-    for (int i = 0; i < branches.size(); ++i)
+    for (size_t i = 0; i < branches.size(); ++i)
     {
         branches[i]->generateIL(instrs, context, destReg);
     }
