+++ src/codegen/allocator.cpp
@@ -7,12 +7,12 @@ void Allocator::allocate(Allocation allocation)
 
 int Allocator::getAllocationOffset(std::string name) const
 {
-    for (int i = _frames.size() - 1; i >= 0; --i)
+    for (size_t i = _frames.size() - 1; i >= 0; --i)
     {
         if (_frames[i].isAllocated(name))
         {
             int offset = _frames[i].getAllocationOffset(name);
-            for (int j = i + 1; j < _frames.size(); ++j) { offset += _frames[j].getFrameSize(); }
+            for (size_t j = i + 1; j < _frames.size(); ++j) { offset += _frames[j].getFrameSize(); }
             return offset;
         }
     }
@@ -22,7 +22,7 @@ int Allocator::getAllocationOffset(std::string name) const
 
 bool Allocator::isAllocated(std::string name) const
 {
-    for (int i = 0; i < _frames.size(); ++i)
+    for (size_t i = 0; i < _frames.size(); ++i)
     {
         if (_frames[i].isAllocated(name)) { return true; }
     }
@@ -43,7 +43,7 @@ void Allocator::pushFrame()
 int Allocator::stackSize() const
 {
     int size = 0;
-    for (int i = 0; i < _frames.size(); ++i)
+    for (size_t i = 0; i < _frames.size(); ++i)
     {
         size += _frames[i].getFrameSize();
     }
