+++ src/context/pythonContext.cpp
@@ -3,7 +3,7 @@
 void PythonContext::indentStream(std::ostream &os, int scopeDepth) const
 {
     os << std::endl;
-    for (int i = 0; i < scopeDepth * PY_INDENT_SIZE; i++)
+    for (size_t i = 0; i < scopeDepth * PY_INDENT_SIZE; i++)
     {
         os << " ";
     }
@@ -15,7 +15,7 @@ void PythonContext::dumpGlobals(std::ostream &os, int scopeDepth) const
     {
         indentStream(os, scopeDepth);
         os << "global ";
-        for (int i = 0; i < _globalIdentifiers.size(); ++i)
+        for (size_t i = 0; i < _globalIdentifiers.size(); ++i)
         {
             if (i > 0)
             {
