+++ src/helpers/utils.cpp
@@ -5,7 +5,7 @@ std::string Utils::padString(std::string str, int maxPadding)
 {
     std::stringstream ss;
     ss << str;
-    for (int i = 0; i < maxPadding - str.size(); i++) { ss << " "; }
+    for (size_t i = 0; i < maxPadding - str.size(); i++) { ss << " "; }
     return ss.str();
 }
